Email reports and SMTP username + identity
Adds basic emails for reports.
New config "emailuser" which is used as the SMTP auth username.
Config "emailaddress" is now used for setting the identity instead of auth.
Emails now show as "From: FChannel"

Note: With this change the identity will no longer be blank because of
the included display name.
If an email isn't set in "emailaddress" then the email will be addressed
"From: FChannel <>", this may cause delivery problems so needs testing.