Add text and flash board types
Board type can be set on board creation or from the board admin page, defaults to imageboard.

Textboards do not display any attachments, local replies to textboard threads should return an error to the poster if they try to include an attachment, remote replies with media are not fully blocked/stripped yet, any replies with an attachment and empty comment are hidden.
The textboard index/thread pages are identical to imageboard, with all attachment stuff ripped out, needs work.

Flashboards threads must have an SWF or FLV file to post, replies can be any of the SupportedFiles types set in config/config.go.
All boards now have a /list view which displays all thread info in a table, this is virtually identical to the flashboard index.